Rename e2e/utils/connection-watcher.js to e2e/utils/helpers.js
